feat: add Staff Performance Tracker service with 77 tests
StaffPerformanceService — tracks and analyzes staff member performance
across rental transactions with composite scoring, leaderboard rankings,
team summaries, and actionable coaching recommendations.

New files:
  Vidly/Models/StaffModels.cs — StaffMember, StaffTransaction,
    StaffTransactionType, StaffPerformanceReport, StaffRankingEntry,
    TeamPerformanceSummary (with ToTextReport())
  Vidly/Services/StaffPerformanceService.cs — staff CRUD, transaction
    recording, individual reports, leaderboards, team summaries,
    period comparison, hourly activity, peak hours, feedback analysis,
    rating distribution, five-star streaks, daily trends
  Vidly.Tests/StaffPerformanceServiceTests.cs — 77 tests covering
    constructor validation, CRUD ops, transaction recording, reports,
    leaderboard, team summary, period comparison, hourly activity,
    feedback filtering, streaks, scoring, grading, edge cases

Also fixed pre-existing build errors:
  - PricingService: renamed TierComparison -> PricingTierComparison
    to resolve ambiguity with Models.TierComparison
  - RevenueAnalyticsService: cast growth to decimal for GrowthPercent
  - MembershipTierServiceTests: fix InMemoryRentalRepository ctor args,
    fix void Add() return assignment
  - RevenueAnalyticsServiceTests: use static Reset() calls, fix
    decimal/double comparison

+++ Vidly/Services/RevenueAnalyticsService.cs
@@ -444,7 +444,7 @@ namespace Vidly.Services
                     Month = m.Month,
                     Revenue = m.Revenue,
                     RentalCount = m.Count,
-                    GrowthPercent = growth,
+                    GrowthPercent = (decimal)growth,
                 });
 
                 prevRevenue = m.Revenue;
